requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
PyMySQL>=1.0
psycopg2-binary
python-dotenv>=1.0
google
google-genai>=0.2
langchain>=0.1
yfinance
websocket-client
supabase>=1.0
APScheduler>=3.9